+++ test/__mocks__/sqlite-wasm-http.js
@@ -2,7 +2,7 @@ module.exports = {
     createSQLiteHTTPPool: () =>
         Promise.resolve({
             open: () => Promise.resolve({}),
-            exec: () => Promise.resolve([{row: ['minZoom', '12']}, {row: ['format', 'png']}]),
+            exec: () => Promise.resolve([{row: ['minzoom', '12']}, {row: ['format', 'png']}]),
             close: () => Promise.resolve(undefined)
         })
 };